// src/api/generated/model/createSportPreferenceDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */
import type { CreateSportPreferenceData } from './createSportPreferenceData';

export interface CreateSportPreferenceDto {
  /** The sport preferences of the user */
  sportPreferences: CreateSportPreferenceData[];
}